==> 84.ADpolypeptide/60.MergeAll/03.Select_signals_LVFA.Rmd <==
---
title: "Signal selecting"
author: "Chao Tang"
date: 'Report created: `r Sys.Date()`'
output: 
  html_document: 
    code_folding: "hide"
    toc: true
    toc_depth: 4
    toc_float: 
      collapsed: false
    number_sections: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, 
                      warning = FALSE, 
                      message = FALSE, 
                      fig.align = 'center')
knitr::opts_knit$set(root.dir = "/mnt/raid61/Personal_data/tangchao/AANanopore")
```

```{r}
library(factoextra)
library(data.table)
library(Biostrings)
library(patchwork)
library(ggplot2)
library(ggrepel)
library(IRanges)
library(dplyr)
library(lsa)
```

```{r}
KnnSelecting <- function(signals, dists, dist_method = "euclidean", k = 20, noise = 0, index = NULL) {
  stopifnot(noise < k)
  if(is(dists, "dist")) {
    knn1 <- dbscan::kNN(dists, k = k)
    knn2 <- apply(knn1$id, 2, function(x) x %in% which(attr(dists, "Labels") %in% grep("^Noise", rownames(knn1$id), value = TRUE)))
  } else {
    if(is(dists, "kNN")) {
      if(k > dists$k) stop("K too big")
      knn1 <- dists
      knn2 <- apply(knn1$id[, 1:k], 2, function(x) x %in% which(grepl("^Noise", rownames(knn1$id))))
    } else {
      knn1 <- dbscan::kNN(dists[[dist_method]], k = k)
      knn2 <- apply(knn1$id, 2, function(x) x %in% which(attr(dists[[dist_method]], "Labels") %in% grep("^Noise", rownames(knn1$id), value = TRUE)))
    }
  }
  
  tu <- row.names(knn1$id)[which(rowSums(knn2) <= noise)]
  # tu <- union(tu, knn2[which(rowSums(knn2) <= noise), 1])
  tu <- grep("^Noise", tu, invert = TRUE, value = TRUE)
  ggplot(signals, aes(x = Blockade, y = DwellTime, colour = !ID %in% tu)) + 
    geom_point(size = .5) + 
    scale_y_log10() + 
    scale_colour_manual(values = c("#E41A1C", "#377EB8")) + 
    theme_classic() + 
    theme(legend.position = "none") -> p1
  ggplot(signals[ID %in% tu], aes(x = Blockade, y = DwellTime, colour = ID %in% tu)) + 
    geom_point(size = .5) + 
    scale_y_log10() + 
    scale_colour_manual(values = "#E41A1C") + 
    theme_classic() + 
    theme(legend.position = "none") + 
    labs(title = length(tu)) -> p2
  ggplot(signals[!ID %in% tu], aes(x = Blockade, y = DwellTime, colour = ID %in% tu)) + 
    geom_point(size = .5) + 
    scale_y_log10() + 
    scale_colour_manual(values = "#377EB8") + 
    theme_classic() + 
    theme(legend.position = "none") -> p3
  if(!is.null(index)) {
    p1 <- p1 + geom_vline(xintercept = index, colour = "red")
    p2 <- p2 + geom_vline(xintercept = index, colour = "red")
    p3 <- p3 + geom_vline(xintercept = index, colour = "red")
  }
  print(p1 + p2 + p3)
  return(tu)
}

RemoveOutlier <- function(tree, n = 10, Keep = 0.95) {
  h <- length(tree$height)
  g <- cutree(tree, h = tree$height[h])
  g <- data.table(tip = names(g), cluster = g)
  
  while (g[, .N, cluster][N >= n, sum(N)/length(tree$labels)] > Keep) {
    h <- h - 1
    g <- cutree(tree, h = tree$height[h])
    g <- data.table(tip = names(g), cluster = g)
  }
  cat(paste0("Keep: ", round(g[, .N, cluster][N >= n, sum(N)/length(tree$labels)] * 100, 2), "%"))
  g[cluster %in% g[, .N, cluster][N >= n, cluster], tip]
}
```

```{r}
AABlockade <- lapply(list.files("./analysis/61.SignalSelecting/01.StandardAA/02.SelectedSignals", "_V2.Rds", full.names = T), function(x) readRDS(x)$Summary)
AABlockade <- do.call(rbind, AABlockade)
AABlockade <- AABlockade[State == "State1"]
```

# LVFA

```{r}
meta0 <- as.data.table(openxlsx::read.xlsx("./analysis/84.ADpolypeptide/60.MergeAll/metainformation.xlsx", sheet = 1))
sig_file <- mapply(meta0[, file_id], FUN = function(x) grep(".MainL0.txt", list.files("./analysis/84.ADpolypeptide", x, full.names = T, recursive = T), value = T))
meta0 <- merge(meta0, data.table(file_id = names(unlist(sig_file)), sig_file = unlist(sig_file)), all.x = TRUE)
meta0 <- meta0[file.exists(sig_file)]
meta0[, unique(amino_acid)]
```

```{r}
meta <- meta0[amino_acid == "LVFA" & file_id != "20231026_0019.1"]
```

```{r fig.width=24, fig.height=6}
dataset <- do.call(rbind, lapply(meta[amino_acid == "LVFA", sig_file], fread))[Blockade < 0.35 & DwellTime > 0.3]
B_file <- paste0("./analysis/81.ABFProcessing/FeatureMatrix/FeatureMatrix_", unique(meta[amino_acid == "LVFA", file_name]), ".txt")
FeatureMatrix <- do.call(rbind, lapply(B_file, fread))
FeatureMatrix <- FeatureMatrix[ID %in% dataset$ID]
FeatureMatrix <- data.frame(FeatureMatrix[, grepl("^X", colnames(FeatureMatrix)) | colnames(FeatureMatrix) %in% c("DwellTime", "DeltaMean", "StageSD", "CurrentSD", "SignalCurrentPercent", "SignalCurrentWidth", "Blockade"), with = F], row.names = FeatureMatrix[[1]])
```

# One by one

## 20231025_0001.1

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/84.ADpolypeptide/60.MergeAll/02.SignalDistance/LVFA_20231025_0001.1_euclidean_distance_knn.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "20231025_0001.1"], dists = alldists, k = 10, noise = 1, index = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LVFA", ""))], Blockade])
```

```{r}
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  geom_vline(xintercept = c(0.24)) + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r}
sigs <- intersect(sigs, dataset[Blockade < 0.24, ID])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(1 - cor(t(subtab), method = "spearman"))
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 100, Keep = 0.8)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```


```{r}
fwrite(dataset[ID %in% sigs], file.path("./analysis/84.ADpolypeptide/60.MergeAll/03.SelectedSignals/LVFA_20231025_0001.1_euclidean_distance_knn_selected.signals.txt"), sep = "\t", quote = F)
```



## 20231025_0002.1

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/84.ADpolypeptide/60.MergeAll/02.SignalDistance/LVFA_20231025_0002.1_euclidean_distance_knn.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "20231025_0002.1"], dists = alldists, k = 10, noise = 1, index = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LVFA", ""))], Blockade])
```

```{r}
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  geom_vline(xintercept = c(0.24)) + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r}
sigs <- intersect(sigs, dataset[Blockade < 0.24, ID])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(1 - cor(t(subtab), method = "spearman"))
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 200, Keep = 0.8)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```


```{r}
fwrite(dataset[ID %in% sigs], file.path("./analysis/84.ADpolypeptide/60.MergeAll/03.SelectedSignals/LVFA_20231025_0002.1_euclidean_distance_knn_selected.signals.txt"), sep = "\t", quote = F)
```



## 20231025_0003.1

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/84.ADpolypeptide/60.MergeAll/02.SignalDistance/LVFA_20231025_0003.1_euclidean_distance_knn.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "20231025_0003.1"], dists = alldists, k = 10, noise = 1, index = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LVFA", ""))], Blockade])
```

```{r}
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  geom_vline(xintercept = c(0.24)) + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r}
sigs <- intersect(sigs, dataset[Blockade < 0.24, ID])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(1 - cor(t(subtab), method = "spearman"))
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 200, Keep = 0.8)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```


```{r}
fwrite(dataset[ID %in% sigs], file.path("./analysis/84.ADpolypeptide/60.MergeAll/03.SelectedSignals/LVFA_20231025_0003.1_euclidean_distance_knn_selected.signals.txt"), sep = "\t", quote = F)
```



## 20231026_0011.1

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/84.ADpolypeptide/60.MergeAll/02.SignalDistance/LVFA_20231026_0011.1_euclidean_distance_knn.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "20231026_0011.1"], dists = alldists, k = 10, noise = 0, index = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LVFA", ""))], Blockade])
```

```{r}
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  geom_vline(xintercept = c(0.24)) + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r}
sigs <- intersect(sigs, dataset[Blockade < 0.24, ID])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(1 - cor(t(subtab), method = "spearman"))
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 200, Keep = 0.8)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```


```{r}
fwrite(dataset[ID %in% sigs], file.path("./analysis/84.ADpolypeptide/60.MergeAll/03.SelectedSignals/LVFA_20231026_0011.1_euclidean_distance_knn_selected.signals.txt"), sep = "\t", quote = F)
```


## 20231026_0012.1

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/84.ADpolypeptide/60.MergeAll/02.SignalDistance/LVFA_20231026_0012.1_euclidean_distance_knn.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "20231026_0012.1"], dists = alldists, k = 10, noise = 0, index = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LVFA", ""))], Blockade])
```

```{r}
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  geom_vline(xintercept = c(0.17, 0.24)) + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r}
sigs <- intersect(sigs, dataset[Blockade > 0.17 & Blockade < 0.24, ID])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(1 - cor(t(subtab), method = "spearman"))
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 200, Keep = 0.8)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```


```{r}
fwrite(dataset[ID %in% sigs], file.path("./analysis/84.ADpolypeptide/60.MergeAll/03.SelectedSignals/LVFA_20231026_0012.1_euclidean_distance_knn_selected.signals.txt"), sep = "\t", quote = F)
```


## 20231026_0013.1

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/84.ADpolypeptide/60.MergeAll/02.SignalDistance/LVFA_20231026_0013.1_euclidean_distance_knn.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "20231026_0013.1"], dists = alldists, k = 10, noise = 0, index = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LVFA", ""))], Blockade])
```

```{r}
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  geom_vline(xintercept = c(0.17, 0.24)) + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r}
sigs <- intersect(sigs, dataset[Blockade > 0.17 & Blockade < 0.24, ID])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(1 - cor(t(subtab), method = "spearman"))
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 200, Keep = 0.8)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```


```{r}
fwrite(dataset[ID %in% sigs], file.path("./analysis/84.ADpolypeptide/60.MergeAll/03.SelectedSignals/LVFA_20231026_0013.1_euclidean_distance_knn_selected.signals.txt"), sep = "\t", quote = F)
```



## 20231026_0016.1

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/84.ADpolypeptide/60.MergeAll/02.SignalDistance/LVFA_20231026_0016.1_euclidean_distance_knn.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "20231026_0016.1"], dists = alldists, k = 10, noise = 0, index = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LVFA", ""))], Blockade])
```

```{r}
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  geom_vline(xintercept = c(0.17, 0.24)) + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r}
sigs <- intersect(sigs, dataset[Blockade > 0.17 & Blockade < 0.24, ID])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(1 - cor(t(subtab), method = "spearman"))
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 10, Keep = 0.98)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```


```{r}
fwrite(dataset[ID %in% sigs], file.path("./analysis/84.ADpolypeptide/60.MergeAll/03.SelectedSignals/LVFA_20231026_0016.1_euclidean_distance_knn_selected.signals.txt"), sep = "\t", quote = F)
```


## 20231026_0017.1

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/84.ADpolypeptide/60.MergeAll/02.SignalDistance/LVFA_20231026_0017.1_euclidean_distance_knn.Rds"))
sigs <- KnnSelecting(signals = dataset[A == "20231026_0017.1"], dists = alldists, k = 10, noise = 0, index = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LVFA", ""))], Blockade])
```

```{r}
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  geom_vline(xintercept = c(0.17, 0.24)) + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r}
sigs <- intersect(sigs, dataset[Blockade > 0.17 & Blockade < 0.24, ID])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(1 - cor(t(subtab), method = "spearman"))
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 10, Keep = 0.98)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```


```{r}
fwrite(dataset[ID %in% sigs], file.path("./analysis/84.ADpolypeptide/60.MergeAll/03.SelectedSignals/LVFA_20231026_0017.1_euclidean_distance_knn_selected.signals.txt"), sep = "\t", quote = F)
```



# All

```{r fig.width=24, fig.height=6}
alldists <- readRDS(file.path("./analysis/84.ADpolypeptide/60.MergeAll/02.SignalDistance/LVFA_spearman_distance_knn.Rds"))
sigs <- KnnSelecting(signals = dataset, dists = alldists, k = 7, noise = 1, index = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LVFA", ""))], Blockade])
```

```{r}
sigs <- intersect(sigs, dataset[Blockade < 0.24 & A %in% c("20230330 _0006.1", "20231018_0003.1"), ID])
```

```{r}
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs)) + 
  geom_vline(xintercept = c(0.16, 0.188))
```

```{r}
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.25)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sigs)
res.ds <- stats::dist(1 - cor(t(subtab), method = "spearman"))
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sigs <- RemoveOutlier(tree = res.hc, n = 10, Keep = 0.98)
ggplot(dataset[ID %in% sigs], aes(x = Blockade, y = DwellTime, colour = ID %in% sigs)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_x_continuous(limits = c(0.1, 0.3)) + 
  scale_colour_manual(values = "#E41A1C") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sigs))
```

```{r fig.width=6, fig.height=4}
ggplot(dataset, aes(x = Blockade, y = DwellTime, colour = !ID %in% sigs)) + 
  geom_point(size = 0.4) + 
  scale_y_log10() + 
  theme_base() + theme(legend.position = "none", plot.background = element_blank())
```

```{r}
fwrite(dataset[ID %in% sigs], file.path("./analysis/84.ADpolypeptide/60.MergeAll/03.SelectedSignals/LVFA_euclidean_distance_knn_selected.signals.txt"), sep = "\t", quote = F)
```
